added jansson to the main project

+++ deps/jansson/src/dump.c
@@ -230,7 +230,7 @@ static int do_dump(const json_t *json, size_t flags, int depth,
                 goto array_error;
             array->visited = 1;
 
-            n = json_array_size(json);
+            n = (int)json_array_size(json);
 
             if(dump("[", 1, data))
                 goto array_error;
